requirements.txt
Flask
python-magic
imutils
numpy
cmake
dlib
opencv-python
matplotlib
pandas
Pillow
scikit-image
